# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Orion — Declarative Services Runtime API",
    "description": "Declarative services runtime platform.\n\n**Authentication.** The admin API (`/api/v1/admin/*`), the Prometheus endpoint (`/metrics`), and the trace read endpoints (`/api/v1/admin/traces*`) require an admin API key when `admin_auth.enabled` is true — the default in the shipped Helm chart and HA compose files. Operations that need it carry a `security` block; see the `admin_bearer` / `admin_api_key` schemes for how the key is presented and for the `admin_auth.header` config key that selects between them.\n\nThe data plane (`/api/v1/data/{channel}`) and the health probes are deliberately unauthenticated: channel-level access control is expressed through each channel's `validation_logic` and CORS configuration.\n\n**Errors.** Every non-2xx response uses the `ErrorResponse` envelope: `{\"error\": {\"code\", \"message\", \"request_id\"}}`, plus `details` on validation failures.",
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/admin/backups": {
      "get": {
        "tags": [
          "Backups"
        ],
        "operationId": "list_backups",
        "responses": {
          "200": {
            "description": "List of backup files in the configured backup directory",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_Vec_BackupListItem"
                }
              }
            }
          },
          "400": {
            "description": "Unavailable in cluster mode — backups are node-local files",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin API key. Only returned when `admin_auth.enabled` is true.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal error (`INTERNAL_ERROR`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_bearer": []
          },
          {
            "admin_api_key": []
          }
        ]
      },
      "post": {
        "tags": [
          "Backups"
        ],
        "operationId": "create_backup",
        "responses": {
          "200": {
            "description": "Backup created (SQLite only — VACUUM INTO a timestamped file)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_BackupFile"
                }
              }
            }
          },
          "400": {
            "description": "Backup unavailable (non-SQLite backend, or cluster mode — use managed-DB snapshots/PITR)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin API key. Only returned when `admin_auth.enabled` is true.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal error (`INTERNAL_ERROR`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_bearer": []
          },
          {
            "admin_api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataEnvelope_BackupFile": {
        "type": "object",
        "description": "The `{\"data\": …}` envelope every admin 2xx carries (R17).\n\nGeneric so each resource gets a described response without a hand-written\nwrapper struct per endpoint. utoipa inlines the type parameter, so\n`DataEnvelope<WorkflowResponse>` publishes the full shape. Before R22, 44\nof the 48 2xx responses had no `content` block at all.",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "description": "The backup `POST /api/v1/admin/backups` just wrote.",
            "required": [
              "filename",
              "path",
              "size_bytes",
              "created_at"
            ],
            "properties": {
              "created_at": {
                "type": "string"
              },
              "filename": {
                "type": "string"
              },
              "path": {
                "type": "string",
                "description": "Absolute path on the node that served the request."
              },
              "size_bytes": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              }
            }
          }
        }
      },
...
      "DataEnvelope_Vec_BackupListItem": {
        "type": "object",
        "description": "The `{\"data\": …}` envelope every admin 2xx carries (R17).\n\nGeneric so each resource gets a described response without a hand-written\nwrapper struct per endpoint. utoipa inlines the type parameter, so\n`DataEnvelope<WorkflowResponse>` publishes the full shape. Before R22, 44\nof the 48 2xx responses had no `content` block at all.",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "One row of `GET /api/v1/admin/backups`. Deliberately not [`BackupFile`]:\nthe listing reads `mtime` off the directory entry and has no `path`, so\nthe two responses differ by two fields.",
              "required": [
                "filename",
                "size_bytes",
                "modified_at"
              ],
              "properties": {
                "filename": {
                  "type": "string"
                },
                "modified_at": {
                  "type": "string"
                },
                "size_bytes": {
                  "type": "integer",
                  "format": "int64",
                  "minimum": 0
                }
              }
            }
          }
        }
      },
...
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code, e.g. `NOT_FOUND`, `VALIDATION_ERROR`,\n`RATE_LIMITED`, `CIRCUIT_OPEN`, `TIMEOUT`, `INTERNAL_ERROR`.",
            "example": "VALIDATION_ERROR"
          },
          "details": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/ErrorFieldDetail"
            },
            "description": "Per-field breakdown, present only on `VALIDATION_ERROR`."
          },
          "message": {
            "type": "string",
            "description": "Human-readable message. Internal 5xx messages are replaced with a\ngeneric string; the detail is logged and kept in the trace."
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Echo of the request's `x-request-id`. Omitted when the request carried\nno id and none was generated."
          }
        }
      },
      "ErrorFieldDetail": {
        "type": "object",
        "description": "One entry of `error.details` on a validation failure.",
        "required": [
          "path",
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "expected": {},
          "got": {},
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string",
            "description": "JSON pointer-ish path to the offending field.",
            "example": "data.amount"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Error response body matching Orion's\n`{\"error\": {\"code\": \"...\", \"message\": \"...\", \"request_id\": \"...\"}}` format.\n\nEvery 4xx/5xx from `OrionError` — and the 500 emitted by the panic-recovery\nlayer — uses this envelope.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
...
    },
    "securitySchemes": {
      "admin_api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Alternative to `admin_bearer`, active when `admin_auth.header` names a header other than `Authorization`. The raw key is sent as that header's value with no `Bearer ` prefix. `X-API-Key` is the conventional choice and is what this document shows, but the header name is deployment-specific — substitute whatever `admin_auth.header` is set to. Exactly one of the two schemes is live at a time; they are listed as alternatives because this document is generated without knowledge of a deployment's config."
      },
      "admin_bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Admin API key presented as `Authorization: Bearer <key>` — the default. Enforced on `/api/v1/admin/*`, `/metrics`, and `/api/v1/admin/traces*` whenever `admin_auth.enabled` is true, which the shipped Helm chart and HA compose files set. Keys come from `admin_auth.api_keys`, either in plaintext or as `sha256:<64-hex>` digests. The data plane (`POST /api/v1/data/{channel}`) is not covered by this scheme."
      }
    }
  },
...
}